Used Car condition prediction
The used car market has significantly grown in recent times, with clients ranging from used car dealers and buyers. Used cars might not be as good as new ones, but when one, running short on a budget, it can be the best option for a while. The Indian used car market is segmented by the organized and unorganized segments. However, C2C (customer to customer) channel has also been used for the sales of pre-owned cars in the market. Also, some service providers are commercially available to guide and providing a platform for the user. There are many factors that involved attention while purchasing the used car.